Add a limit to counter values.
When counters get too big it becomes impossible to fix them.

We had a counter with a shard with a value of 1.7976931348623159e+308
while another shard had the value -1.7976931348623159e+308.
Since these are both the max and min of a double it is impossible to
correct this using incrby commands.